now you can not follow yourself

+++ app/Http/Livewire/FollowUser.php
@@ -15,14 +15,16 @@ class FollowUser extends Component
     }
     public function follow()
     {
-        $follow = DB::table('follows')->where([['follower_id', auth()->id()], ['followed_id', $this->followed_id]])->first();
-        if ($follow) {
-            Follow::where('id', $follow->id)->delete();
-        } else {
-            Follow::create([
-                'followed_id' => $this->followed_id,
-                'follower_id' => auth()->id(),
-            ]);
+        if (auth()->id() != $this->followed_id) {
+            $follow = DB::table('follows')->where([['follower_id', auth()->id()], ['followed_id', $this->followed_id]])->first();
+            if ($follow) {
+                Follow::where('id', $follow->id)->delete();
+            } else {
+                Follow::create([
+                    'followed_id' => $this->followed_id,
+                    'follower_id' => auth()->id(),
+                ]);
+            }
         }
     }
     public function render()
